pyproject.toml
[project]
name = "resume-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "unstructured[docx]>=0.18.24",
]
